[fix][broker] Fix deleting topic not delete the related topic policy and schema. (#21093) (#38)
* [fix][broker] Fix deleting topic not delete the related topic policy and schema. (#21093)

Fixes #21075

### Motivation

When the topic is loaded, it will delete the topic-level policy if it is enabled. But if the topic is not loaded, it will directly delete through managed ledger factory. But then we will leave the topic policy there. When the topic is created next time, it will use the old topic policy

### Modifications

When deleting the topic, delete the schema and topic policies even if the topic is not loaded.

(cherry picked from commit a1405ea006f175b1bd0b9d28b9444d592fb4a010)
Signed-off-by: Zixuan Liu <nodeces@gmail.com>

* Fix test

Signed-off-by: Zixuan Liu <nodeces@gmail.com>

---------

Signed-off-by: Zixuan Liu <nodeces@gmail.com>
Co-authored-by: GuoJiwei <technoboy@apache.org>